requirements.txt
coinbase-agentkit>=0.1.0
langchain>=0.1.0
openai>=1.0.0
click>=8.0.0
python-dotenv>=0.19.0
web3>=6.0.0
aiohttp>=3.8.0
pyyaml>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
pylint>=2.15.0 